+++ R/ggcoefstats.R
@@ -110,10 +110,6 @@
 #' @param stats.label.args Additional arguments that will be passed to
 #'   `ggrepel::geom_label_repel` geom. Please see documentation for that
 #'   function to know more about these arguments.
-#' @param package Name of package from which the palette is desired as string
-#' or symbol.
-#' @param palette Name of palette as string or symbol.
-#' @param direction Either `1` or `-1`. If `-1` the palette will be reversed.
 #' @param only.significant If `TRUE`, only stats labels for significant effects
 #'   is shown (Default: `FALSE`). This can be helpful when a large number of
 #'   regression coefficients are to be displayed in a single plot. Relevant only